refactor: 코드 위치 변경
